+++ cache/test/cache/registry_download_events_pipeline_test.exs
@@ -108,6 +108,7 @@ defmodule Cache.RegistryDownloadEventsPipelineTest do
         headers = options[:headers]
         assert {"content-type", "application/json"} in headers
         assert Enum.any?(headers, fn {key, _value} -> key == "x-cache-signature" end)
+        assert Enum.any?(headers, fn {key, _value} -> key == "x-cache-endpoint" end)
 
         {:ok, %{status: 202, body: ""}}
       end)