+++ server/lib/tuist/gradle/build.ex
@@ -38,6 +38,8 @@ defmodule Tuist.Gradle.Build do
     field :tasks_skipped_count, Ch, type: "UInt32"
     field :tasks_no_source_count, Ch, type: "UInt32"
     field :cacheable_tasks_count, Ch, type: "UInt32"
+    field :requested_tasks, {:array, Ch}, type: "String", default: []
     field :inserted_at, Ch, type: "DateTime"
+    has_many :machine_metrics, Tuist.Builds.BuildMachineMetric, foreign_key: :gradle_build_id, references: :id
   end
 end